requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
pytest
httpx
google-cloud-storage
pydantic
opentelemetry-exporter-gcp-trace
